adds working sample api
