+++ readrr_api/__init__.py
@@ -1,12 +1,12 @@
 import os
 
 from flask import Flask
-from readerr_api.routes.recommendations import recommendations
+from .routes.recommendations import recommendations
 
 
 def create_app():
     app = Flask(__name__)
 
-    app.register_blueprint(primary_routes)
+    app.register_blueprint(recommendations)
 
     return app
